requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
python-dotenv==1.0.1
openai>=1.0.0
numpy>=1.26.0
scikit-learn>=1.5.0
